:sparkles: Support CommonJS modules
