+++ src/index.ts
@@ -1,22 +1,3 @@
-import Config, { Middleware } from './Config';
-import ConfigSource from './ConfigSource';
-import entries from './entries';
-import placeholder from './middlewares/placeholder';
+import nc from './nc';
 
-export default function config(...args: any[]) {
-  const middlewares: Middleware[] = [
-    placeholder(),
-  ];
-
-  return {
-    use(middleware: Middleware): void {
-      middlewares.push(middleware);
-    },
-
-    defaults<T>(obj: T): T {
-      const sources = [...args, obj].map(x => new ConfigSource(entries(x)));
-      const config = new Config(sources, middlewares);
-      return config.extend(obj);
-    },
-  };
-}
+export = Object.assign(nc, { default: nc });
